+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Support for computed members map variable
+
+### Fixed
+
+- Complete test suite
+
 ## [0.0.3]
 
 ### Fixed